refactor: (#152) UI에 관련된 상수, 타입들을 분리
