+++ frontend/src/types/comment.ts
@@ -1,5 +1,3 @@
-import { COMMENT_ACTION, COMMENT_USER } from '@constants/comment';
-
 export interface Comment {
   id: number;
   member: {
@@ -20,16 +18,4 @@ export interface CommentResponse {
   content: string;
   createdAt: string;
   updatedAt: string;
-}
-
-export type CommentAction = (typeof COMMENT_ACTION)[keyof typeof COMMENT_ACTION];
-
-export type CommentUser = (typeof COMMENT_USER)[keyof typeof COMMENT_USER];
-
-export type CommentMenu = Exclude<CommentUser, 'GUEST'>;
-
-export interface CommentMenuItem {
-  content: string;
-  color: 'black' | 'red';
-  action: CommentAction;
 }
